requirements.txt
Django==1.11.17
django-environ==0.4.4
mysqlclient==1.3.12
django-extensions==1.9.9
django-oscar==1.6.4
psycopg2-binary==2.7.6.1
Pillow==5.3.0
